Add ZSteam un-chunking notes
For performance reasons it is important to know which `ZStream` operators destroy the chunking structure. In this PR we attempt to find all those operators and add a note about this behavior.

The note no longer recommends to use `rechunk` because in some contexts this is spectacular bad advice (for example, rechunking a stream that only produces items once every few minutes, common in zio-kafka applications, causes unpredictable high processing delays).